Update checkout.py
Removed default location actually being applied, for now, its code can stay IMO.  Need to consider this, but I don't see an actual production need for a default location.  What actually needs to happen is to query snipe for the locations in the system and then present that as the choices to the user, ideally I don't want to hard code any of that.

The Default model can stay for now, as i mainly expect to be using this for AP535s now, but same thing, I expect to update this to pull the models from your snipe instance, and provide them as a list of choices.