[all] graceful CLI error handling
added an error-handling wrapper and associated plumbing such that when an app crashes, by default it no longer prints out a huge annoying stack trace, unless the user has passed in `--debug` or `--trace`

We often times use python's exception handling system to return errors to the user in situations where the program cannot continue. This is great, because it means consistent error handling behaviour wether a user is importing one of our tools as a python library, or using one of our cli tools; and it means we can troubleshoot errors when things unexpectedly go awry.

It's not so great when an end user of our tools makes a mistake, enters wrong input, and get back a useful error message buried at the end of a (to them) meaningless stack trace. (also it gives them the impression our programs are constantly crashing / unstable, even when that *isn't* true :P)

with these new cli wrappers on all our cli tools, errors will print to screen in a nice, pretty way, unless the user passes `--debug` or `--trace`, in which case the stack trace will be printed to screen as usual