[dev] switch build backend from hatchling to pdm-backend
As part of an ongoing effort to improve package management in the monorepo, we're switching from hatchling to pdm-backend. The biggest reason is that hatchling does not support passing in config at runtime via environment variables or pep517 config-settings.
In order to satisfy all the requirements of issue #29, we have build hooks that run whenever one of our projects gets installed or built or put into a lock file. these build hooks rewrite dependency metadata to satisfy different requirements. without the ability to pass in a goal or target to the build hooks, the hatchling-based ones perform some implicit "magic" to rewrite dependencies based on whatever environment markers they're able to sus out of the build process. I've never really been a big fan of that, I prefer explicit behaviour rather than implicit behaviour. One of the biggest issues with the implicit build hooks is that they would fail to perform as intend when building wheel files, and would produce wheel files that were broken

The new pdm-based build hooks work much better and are much easier to control. The only downside is that in order for pip-installing a folder in the monorepo to "just work", you have to add `--config-settings 'dependencies=local'` to the pip command. it no longer "just works" without that cli flag. :(

On the plus side, this new approach will allow us to create "local wheels" for distribution, ie a wheel file whose monorepo dependencies are pulled from adjacent wheel files. (ex: a bundle of .whl files for all projects in the monorepo, and if you `pip install uoft_aruba.*.whl`, pip will automatically install its dependency `uoft_core` from `./uoft_core.*.whl`)